Additional restructuring, updated .gitignore, added requirements.txt
