+++ requirements.txt
@@ -1,0 +1,7 @@
+docutils
+pygments
+pypiwin32
+kivy_deps.sdl2==0.1.22
+kivy_deps.glew==0.1.12
+kivy_deps.gstreamer==0.1.17
+kivy==1.11.1
