--- src/__init__.py
@@ -1,2 +1,0 @@
-from board import *
-from minimax import *
