+++ src/minimax.py
@@ -4,24 +4,26 @@ from math import inf as max_score
 LENGTH = 3
 SYMBOLS = {'computer': 'O', 'human': 'X', 'empty': ''}
 
-class SimpleBoard():
+
+class SimpleBoard:
     def __init__(self, board):
-        self.board = [[button.text for button in row] for row in board]
-    
+        self.__board = [[button.text for button in row] for row in board]
+
     def __getitem__(self, index):
-        return self.board[index]
+        return self.__board[index]
 
     def __len__(self):
-        return len(self.board)
-    
+        return len(self.__board)
+
     def __iter__(self):
-        return iter(self.board)
-    
+        return iter(self.__board)
+
     def is_full(self):
-        return not any([symbol == SYMBOLS['empty'] for row in self.board for symbol in row])
-    
+        return not any([symbol == SYMBOLS['empty'] for row in self.__board for symbol in row])
+
     def has_won(self):
         return abs(evaluate(self)) == max_score
+
 
 def get_possibilities(board, symbol):
     """
@@ -40,6 +42,7 @@ def get_possibilities(board, symbol):
                 out.append((option, (i, j)))
     return out
 
+
 def evaluate(board):
     """
     :param board:   The board to evaluate
@@ -49,13 +52,14 @@ def evaluate(board):
     two_in_row = [0, 0]
     for line in lines:
         for i in range(len(line)):
-            if line[i] == LENGTH: 
+            if line[i] == LENGTH:
                 return max_score * (-1 if i == 1 else 1)
             if line[i] == LENGTH - 1 and line[1 - i] == 0:
                 two_in_row[i] += 1
     comp_score = 10 ** two_in_row[0] if two_in_row[0] > 0 else 0
     player_score = 2 * (10 ** two_in_row[1]) if two_in_row[1] > 0 else 0
     return comp_score - player_score
+
 
 def check_rows(board):
     """
@@ -67,6 +71,7 @@ def check_rows(board):
         out.append((row.count(SYMBOLS['computer']), row.count(SYMBOLS['human'])))
     return out
 
+
 def check_cols(board):
     """
     :param board:   The game board
@@ -75,14 +80,16 @@ def check_cols(board):
     transpose = [[row[i] for row in board] for i in range(LENGTH)]
     return check_rows(transpose)
 
+
 def check_diags(board):
     """
     :param board:   The game board
     :return:        A list containing how many of each symbol is in each diagonal in :board:
     """
     diagonals = [[board[i][i] for i in range(len(board))],
-                    [board[i][LENGTH - i - 1] for i in range(len(board))]]
+                 [board[i][LENGTH - i - 1] for i in range(len(board))]]
     return check_rows(diagonals)
+
 
 def minimax(board, depth):
     """
@@ -95,7 +102,8 @@ def minimax(board, depth):
     if depth <= 0:
         options = get_possibilities(board, SYMBOLS['computer'])
         return pick_highest(options)
-    return play(board, 'computer', alpha, beta, depth, depth)
+    return make_move(board, 'computer', alpha, beta, depth, depth)
+
 
 def pick_highest(options):
     """
@@ -105,7 +113,8 @@ def pick_highest(options):
     scores = [evaluate(x[0]) for x in options]
     return options[scores.index(max(scores))][1]
 
-def play(board, player, alpha, beta, depth, idepth):
+
+def make_move(board, player, alpha, beta, depth, idepth):
     """
     :param board:   A simplified version of the current board
     :param player:  The player the algorithm is playing as (Can only be a key from SYMBOLS)
@@ -122,9 +131,9 @@ def play(board, player, alpha, beta, depth, idepth):
     options = get_possibilities(board, SYMBOLS[player])
     n_player = 'computer' if player == 'human' else 'human'
     best_index = options[0][1]
-    best_score = play(options[0][0], n_player, alpha, beta, depth - 1, idepth)
+    best_score = make_move(options[0][0], n_player, alpha, beta, depth - 1, idepth)
     for option in options[1:]:
-        score = play(option[0], n_player, alpha, beta, depth - 1, idepth)
+        score = make_move(option[0], n_player, alpha, beta, depth - 1, idepth)
         if better_move(player, score, best_score):
             best_index = option[1]
             best_score = score
@@ -135,6 +144,7 @@ def play(board, player, alpha, beta, depth, idepth):
         if beta <= alpha:
             break
     return best_score if depth != idepth else best_index
+
 
 def better_move(player, score, best_score):
     """
